@@ -9,14 +9,19 @@
 		}
 	],
 	"require": {
-		"php": ">=5.5.0"
+		"php": ">=7.1"
 	},
 	"require-dev": {
-		"phpunit/phpunit": "^5.3.0"
+		"phpunit/phpunit": "*"
 	},
 	"autoload": {
 		"psr-4": {
 			"PNixx\\Crontab\\": "src/"
+		}
+	},
+	"autoload-dev": {
+		"psr-4": {
+			"PNixx\\Crontab\\": "test/"
 		}
 	}
 }
